+++ src/main/java/org/cts/parser/prj/PrjParser.java
@@ -50,6 +50,28 @@ public class PrjParser {
     }
 
     /**
+     * Parses a WKT PRJ String into a {@code PrjElement}.
+     *
+     * <p>
+     * This is the main entry point of the parser.
+     *
+     * @param prjString a WKT string
+     * @return a {@code PrjElement}
+     * @throws PrjParserException if the PRJ cannot be parsed into a CRS for any
+     *                            reason
+     */
+    public PrjElement getAsPrjElement(String prjString) {
+        CharBuffer s = CharBuffer.wrap(prjString);
+        PrjElement e;
+        try {
+            e = parseNode(s);
+        } catch (BufferUnderflowException ex) {
+            throw new PrjParserException("Failed to read PRJ.", ex);
+        }
+        return e;
+    }
+
+    /**
      * Parses a WKT PRJ String into a set of parameters.
      * <p>
      * This is the main entry point of the parser.
@@ -60,14 +82,7 @@ public class PrjParser {
      *                            reason
      */
     public Map<String, String> getParameters(String prjString) {
-        CharBuffer s = CharBuffer.wrap(prjString);
-        PrjElement e;
-        try {
-            e = parseNode(s);
-        } catch (BufferUnderflowException ex) {
-            throw new PrjParserException("Failed to read PRJ.", ex);
-        }
-        return PrjMatcher.match(e);
+        return PrjMatcher.match(getAsPrjElement(prjString));
     }
 
     /**