+++ docs/CHANGELOG.md
@@ -4,3 +4,4 @@
 - Remove unnecessary logger
 - Update github actions
 - Change to central portal publishing
+- Add a new method to return the elements from a PRJ string